refactor(v211): rename source + test files to canonical names
Source:
  humanities_essay_scaffold.py → humanities_essay.py
  (back-compat shim left at old path; removed v2.2.0)

  The public tool name `tool_humanities_essay_scaffold` and the
  function `scaffold_humanities_essay` are unchanged — only the
  module path is renamed. Imports from the old module path resolve
  through the 2-line shim.

Tests (11 files): drop `_v2` suffix since the audit consolidation
landed in v1.6.1 and the test names no longer need to distinguish
from a "v1" of those tests. No collisions; rename is mechanical.
  test_audit_audit_completeness_v2.py → ...completeness.py
  test_audit_audit_synthesis_v2.py    → ...synthesis.py
  test_audit_audit_master_v2.py       → ...master.py
  test_audit_audit_prose_v2.py        → ...prose.py
  test_audit_audit_claims_v2.py       → ...claims.py
  test_audit_audit_reviewer_responses_v2.py → ...reviewer_responses.py
  test_audit_audit_cross_deliverable_consistency_v2.py → ...cross_deliverable_consistency.py
  test_audit_audit_code_quality_v2.py → ...code_quality.py
  test_audit_audit_step_literature_v2.py → ...step_literature.py
  test_audit_audit_figure_coverage_v2.py → ...figure_coverage.py
  test_router_output_v2.py            → test_router_output.py

dashboard_v2*.py shims (created in v2.1.0) intentionally LEFT — they
are scheduled for v2.2.0 MINOR removal per the migration table; PATCH
releases don't remove API surfaces, even shimmed ones.

Test gate: preflight 24/24 · pytest 1605 pass / 13 skip · ruff clean.
