+++ tests/unit/test_audit_audit_figure_coverage.py
